requirements.txt
alembic==1.6.5
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dominate==2.6.0
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
Pillow==8.2.0
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.18
urllib3==1.26.5
visitor==0.1.3
Werkzeug==2.0.1
WTForms==2.3.3
